requirements.txt
openvino-runtime==2024.6.0
Pillow
kivy